Added Bug/Issue Tracker information.
The current bug/issue tracker information has been updated.

Trac does not seem to allow easy adding of a link to the specific component.

GitHub will link to the current issues in the current repository.  The "components" of other issue trackers seem to link here.

Bugzilla links to the component quite nicely.